input,outputの追加

+++ get_e.f90
@@ -1,11 +1,11 @@
 program get_e
 	implicit none
 	integer::i=0
-	real::e=0,a=1.
+	real(8)::e=0.0d0,a=1.0d0
 	do
 		if (1/a<1.0e-10) exit
 		i=i+1
-		e=e+1/a
+		e=e+1.0d0/a
 		a=a*i
 	end do
 	print*,"e =",e
